WebDAV: Refactored filehandling because of changes in trunk
The return value of the function 'obj_file->getFile()' changed after an a update like 'obj_file->update()' form './file.xyz' to './001/file.xyz'. But the file is still stored in the first described location. I fixed this with a refactor and some dirty tricks. This issue is fixed, but MUST be refactored after a discussion with other developers.